requirements.txt
numpy==1.26.2
pandas==2.1.4
web3==5.25.0
matplotlib==3.8.2
scikit-learn==1.3.2
pika==1.3.2
PyYAML==6.0.1
seaborn==0.13.0
# extra added
python-dotenv==1.0.0
